Solves Day 2 parts 1 and 2
Implements solutions for both parts of the Day 2 challenge.

Part 1 calculates the checksum by finding the difference between the maximum and minimum values in each row.

Part 2 calculates the checksum by finding the sum of the results of the divisions of the two numbers that evenly divide each other in each row.